+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 Added
+
+Updated
+
+Fixed
+
+## v3.5.0
+Added
 - timeout property for Run objects
 - data_ids property for Run objects